fix(calendar): bring back search query that I broke
